requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.10.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
